Create node in graph page
